requirements.txt
#dj-database-url==0.4.2
#dj-static==0.0.6
Django==2.0
django-autoslug==1.9.3
django-autoslug-field==0.2.3
django-wkhtmltopdf==3.1.0
djangorestframework==3.7.7
djangorestframework-expander==0.2.3
olefile==0.44
Pillow==4.3.0
python-decouple==3.1
pytz==2017.3
static3==0.7.0
-r requirements-dev.txt
gunicorn
psycopg2
